build: use common package in frontend
