+++ packages/backend/src/submission/controller.ts
@@ -1,5 +1,9 @@
 import type { Request, Response } from 'express'
-import { InvitationDto, SubmissionDto } from '@ssi-trust-registry/common'
+import {
+  FieldError,
+  InvitationDto,
+  SubmissionDto,
+} from '@ssi-trust-registry/common'
 import { createLogger } from '../logger'
 import { RequestWithToken } from '../auth/middleware'
 import partial from 'lodash.partial'
@@ -7,7 +11,6 @@ import { ValidationService } from '../entity/validationService'
 import { SubmissionService } from './service'
 import { EmailClient } from '../email/client'
 import { config } from '../config'
-import { FieldError } from '../errors'
 
 const logger = createLogger(__filename)
