+++ packages/backend/src/submission/service.ts
@@ -2,17 +2,17 @@ import partial from 'lodash.partial'
 import { extendZodWithOpenApi } from '@asteasolutions/zod-to-openapi'
 import { z } from 'zod'
 import { v4 as uuidv4 } from 'uuid'
-import { createLogger } from '../logger'
-import { EntityRepository } from '../entity/service'
-import { FieldError } from '../errors'
-import { createId } from '@paralleldrive/cuid2'
 import {
   Entity,
+  FieldError,
   Invitation,
   InvitationDto,
   Submission,
   SubmissionDto,
 } from '@ssi-trust-registry/common'
+import { createId } from '@paralleldrive/cuid2'
+import { createLogger } from '../logger'
+import { EntityRepository } from '../entity/service'
 
 const logger = createLogger(__filename)
 extendZodWithOpenApi(z)
