+++ packages/common/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@asteasolutions/zod-to-openapi": "^6.2.0",
+    "make-error": "^1.3.6",
     "zod": "^3.22.3"
   }
 }
