+++ packages/common/src/index.ts
@@ -2,3 +2,4 @@ export * from './domain/entity'
 export * from './domain/invitation'
 export * from './domain/schema'
 export * from './domain/submission'
+export * from './errors'
