+++ packages/frontend/package.json
@@ -13,8 +13,8 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^3.3.2",
+    "@ssi-trust-registry/common": "1.0.0",
     "dayjs": "^1.11.10",
-    "make-error": "^1.3.6",
     "next": "^14.0.4",
     "next-runtime-env": "^3.2.0",
     "react": "^18.2.0",
