+++ packages/frontend/src/app/[id]/page.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { Entity } from '@/common/interfaces'
+import Image from 'next/image'
+import { Entity } from '@ssi-trust-registry/common'
 import { backendUrl, betterFetch } from '@/api'
 import { NavigationBreadcrumbs } from '@/common/components/navigation/Breadcrumbs'
-import Image from 'next/image'
 import { Text2xlBold, TextSm, TextSmBold } from '@/common/components/Typography'
 import { Card, CardWrapper } from '@/common/components/Card'
 import { PageContainer } from '@/common/components/PageContainer'
