+++ packages/frontend/src/app/submit/[invitationId]/components/SubmissionForm.tsx
@@ -3,9 +3,9 @@ import React from 'react'
 import { useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { z } from 'zod'
-import { FieldError, backendUrl, betterFetch } from '@/api'
+import { FieldError, Invitation } from '@ssi-trust-registry/common'
+import { backendUrl, betterFetch } from '@/api'
 import { TextInput } from '@/common/components/TextInput'
-import { Invitation } from '@/common/interfaces'
 import { TextArea } from '@/common/components/TextArea'
 import { Checkbox } from '@/common/components/Checkbox'
 import Success from '@/common/assets/Success.svg'
